refactor: organize chat tool output helpers
