+++ apps/dashboard/src/components/ai/chat-tool-block/mcp/constants.ts
@@ -1,0 +1,4 @@
+export const MCP_TOOL_NAME_REGEX = /^mcp_/;
+export const UNDERSCORE_REGEX = /_+/g;
+export const NON_ALPHANUMERIC_WORD_SEPARATOR_REGEX = /[^a-zA-Z0-9]+/g;
+export const TITLE_CASE_WORD_REGEX = /\b\w/g;
