+++ apps/dashboard/src/components/ai/chat-tool-block/tool-output-images/types.ts
@@ -1,0 +1,5 @@
+export interface ToolOutputImage {
+  url: string;
+  mediaType: string;
+  filename?: string;
+}
