+++ apps/dashboard/src/components/ai/chat-tool-block/types.ts
@@ -1,0 +1,23 @@
+export interface ToolCopy {
+  verbs: readonly [present: string, past: string];
+  noun: string;
+  subtitle?: (params: {
+    input: unknown;
+    output: unknown;
+    isStreaming: boolean;
+    isError: boolean;
+  }) => string | undefined;
+  suffix?: (input: unknown, output: unknown) => string | undefined;
+}
+
+export interface ChatToolBlockProps {
+  toolName: string;
+  state: string;
+  input?: unknown;
+  output?: unknown;
+  onApprove?: () => void;
+  onDeny?: () => void;
+  isMcp?: boolean;
+  iconUrl?: string;
+  toolMetadata?: unknown;
+}
